Implement JMS Receiver (Message Driven Bean) to receive client messages

+++ UrbanTrafficSimulator/core/src/main/java/lk/oxo/urbantraffic/model/TrafficData.java
@@ -28,7 +28,7 @@ public class TrafficData implements Serializable {
         this.trafficZone = trafficZone;
     }
 
-    public double getVehicleSpeed() {
+    public long getVehicleSpeed() {
         return vehicleSpeed;
     }
